# By Magdalena (2) and Sebastian Nalepka (2)
# Via Sebastian Nalepka
* 'master' of https://github.com/AnnRe/Aqua_proj:
  fix constans
  - Edycja usera - Powiadomienia mail (mailepassy.txt) - Okno Admina edycja - edycja bazy info seba.txt - fix usuń usera - dodanie funkcji update Pracownik, Umowa, - uzupelnienie constants
  Dodanie przykładowej notatki do bazy
  Edycja i uzupełnienie tabeli Otwarcie_stanowiska

Conflicts:
	aquadrom/DB/DBAdapter.cs

+++ aquadrom/AdminPanel.cs
@@ -28,13 +28,15 @@ namespace aquadrom
 
         public void AdminPanel_Load(object sender, EventArgs e)
         {
-            DataTable dtlista = connector.Select("* from "+Constants.TabPracownik+" p,"+Constants.TabUmowa+" u where p."+Constants.PracownikIDUmowy+"=u."+Constants.UmowaIDu);
+            DataTable dtlista = connector.Select("* from "+Constants.TabPracownik+" p,"+Constants.TabUmowa+" u where p."+Constants.PracownikIDUmowyKol+"=u."+Constants.UmowaIDu);
             dataGridView1.AutoSizeColumnsMode = DataGridViewAutoSizeColumnsMode.AllCells;
             dtlista.Columns.Remove(Constants.UmowaIDu);
-            dtlista.Columns.Remove(Constants.PracownikIDUmowy);
-            dtlista.Columns.Remove(Constants.PracownikHaslo);
+            dtlista.Columns.Remove(Constants.PracownikIDUmowyKol);
+            dtlista.Columns.Remove(Constants.PracownikHasloKol);
             dataGridView1.DataSource = dtlista.DefaultView;
-            dataGridView1.Columns[Constants.PracownikOstrzezenie].Visible = false;
+            dataGridView1.Columns[Constants.PracownikOstrzezenieBadaniaKol].Visible = false;
+            dataGridView1.Columns[Constants.PracownikOstrzezenieKPPKol].Visible = false;
+            dataGridView1.Columns[Constants.PracownikOstrzezenieUmowaKol].Visible = false;
             ColorCheckUser();
         }
 
@@ -45,7 +47,7 @@ namespace aquadrom
 
         private void UsuńToolStripMenuItem_Click(object sender, EventArgs e)
         {
-            if( (DeleteWorker.exist==false) && (EditWorkerWhich.exist==false) )  // jeśli okno DeleteWorker,WhichWorker zamknięte to je otwórz else nic
+            if( (DeleteWorker.exist==false) && (EditWorkerWhich.exist==false) && (EditWorker.exist==false) )  // jeśli okno DeleteWorker,WhichWorker,EditWorker zamknięte to je otwórz else nic
             {
                 DeleteWorker DelWor = new DeleteWorker(this);
                 DelWor.Show();
@@ -66,40 +68,56 @@ namespace aquadrom
             }
         }
 
+        private void DodajUżytkownikówToolStripMenuItem_Click(object sender, EventArgs e)
+        {
+            Form2 AddWor = new Form2();
+            AddWor.Show();
+        }
+
         private void ColorCheckUser()
         {
             bool changeKPP = false;
             bool changemedi = false;
+            bool changeconcract = false;
             DateTime today = DateTime.Today.AddDays(7); // jaki okres przed
 
-            for (int i = 0; i < dataGridView1.ColumnCount; i++)
+            for (int i = 0; i < dataGridView1.RowCount; i++)
             {
                 changeKPP = false;
                 changemedi = false;
-                DateTime KPPdate = DateTime.Parse(dataGridView1.Rows[i].Cells[Constants.PracownikWaznKPP].Value.ToString());
-                DateTime medicaldate = DateTime.Parse(dataGridView1.Rows[i].Cells[Constants.PracownikDataBadan].Value.ToString());
+                changeconcract = false;
 
+                DateTime KPPdate;   // if user to KZ (brak KPPdate) to ustala odległą w przyszłość
+                if(dataGridView1.Rows[i].Cells[Constants.PracownikWaznKPPKol].Value.ToString() == "")
+                    KPPdate = DateTime.Now.AddYears(99);
+
+                else KPPdate = DateTime.Parse(dataGridView1.Rows[i].Cells[Constants.PracownikWaznKPPKol].Value.ToString());
+                DateTime medicaldate = DateTime.Parse(dataGridView1.Rows[i].Cells[Constants.PracownikDataBadanKol].Value.ToString());
+                DateTime concractdate = DateTime.Parse(dataGridView1.Rows[i].Cells[Constants.UmowaKoniecUmowy].Value.ToString());
 
                 if (today >= KPPdate)
                 {
-                    dataGridView1.Rows[i].Cells[Constants.PracownikID].Style.BackColor = Color.Red;
-                    dataGridView1.Rows[i].Cells[Constants.PracownikWaznKPP].Style.BackColor = Color.Red;
+                    dataGridView1.Rows[i].Cells[Constants.PracownikIDpKol].Style.BackColor = Color.Red;
+                    dataGridView1.Rows[i].Cells[Constants.PracownikWaznKPPKol].Style.BackColor = Color.Red;
                     changeKPP = true;
                 }
-
                 if (today >= medicaldate)
                 {
-                    dataGridView1.Rows[i].Cells[Constants.PracownikID].Style.BackColor = Color.Red;
-                    dataGridView1.Rows[i].Cells[Constants.PracownikDataBadan].Style.BackColor = Color.Red;
+                    dataGridView1.Rows[i].Cells[Constants.PracownikIDpKol].Style.BackColor = Color.Red;
+                    dataGridView1.Rows[i].Cells[Constants.PracownikDataBadanKol].Style.BackColor = Color.Red;
                     changemedi = true;
                 }
-
-                if ( (changeKPP == true) || (changemedi == true) ) 
+                if (today >= concractdate)
                 {
-                    sendmail(i,KPPdate,medicaldate,changeKPP);
+                    dataGridView1.Rows[i].Cells[Constants.PracownikIDpKol].Style.BackColor = Color.Red;
+                    dataGridView1.Rows[i].Cells[Constants.UmowaKoniecUmowy].Style.BackColor = Color.Red;
+                    changeconcract = true;
                 }
-                else if (dataGridView1.Rows[i].Cells[Constants.PracownikOstrzezenie].Value.ToString() == "t")
-                    adapter.Update(Constants.TabPracownik + " set " + Constants.PracownikOstrzezenie + "='f' where " + Constants.PracownikID + "=" + dataGridView1.Rows[i].Cells[Constants.PracownikID].Value.ToString());
+
+                if ( (changeKPP == true) || (changeconcract == true) || (changemedi == true) )
+                {
+                  //  sendmail(i, KPPdate, changeKPP, medicaldate, changemedi, concractdate, changeconcract);
+                }
             }
         }
 
@@ -108,20 +126,32 @@ namespace aquadrom
             dataGridView1.ClearSelection();
         }
 
-        private void sendmail(int iterator, DateTime KPPdate, DateTime medicaldate, bool changeKPP)
+        private void sendmail(int iterator, DateTime KPPdate, bool changeKPP, DateTime medicaldate, bool changemedi, DateTime concractdate, bool changeconcract )
         {
             string what = "";
-            if (dataGridView1.Rows[iterator].Cells[Constants.PracownikOstrzezenie].Value.ToString() == "f")
+
+            if ((changeKPP == true) && (dataGridView1.Rows[iterator].Cells[Constants.PracownikOstrzezenieKPPKol].Value.ToString() == "f"))
+            {
+                what += "Data ważności KPP: " + KPPdate.ToString("dd-MM-yyyy") + "\n";
+                adapter.Update(Constants.TabPracownik + " set " + Constants.PracownikOstrzezenieKPPKol + "='t' where " + Constants.PracownikIDpKol + "=" + dataGridView1.Rows[iterator].Cells[Constants.PracownikIDpKol].Value.ToString());
+            }
+            else if ((changemedi == true) && (dataGridView1.Rows[iterator].Cells[Constants.PracownikOstrzezenieBadaniaKol].Value.ToString() == "f"))
+            {
+                what += "Data ważności badań: " + medicaldate.ToString("dd-MM-yyyy") + "\n";
+                adapter.Update(Constants.TabPracownik + " set " + Constants.PracownikOstrzezenieBadaniaKol + "='t' where " + Constants.PracownikIDpKol + "=" + dataGridView1.Rows[iterator].Cells[Constants.PracownikIDpKol].Value.ToString());
+            }
+            else if ((changeconcract == true) && (dataGridView1.Rows[iterator].Cells[Constants.PracownikOstrzezenieUmowaKol].Value.ToString() == "f"))
+            {
+                what += "Data ważności umowy: " + medicaldate.ToString("dd-MM-yyyy") + "\n";
+                adapter.Update(Constants.TabPracownik + " set " + Constants.PracownikOstrzezenieUmowaKol + "='t' where " + Constants.PracownikIDpKol + "=" + dataGridView1.Rows[iterator].Cells[Constants.PracownikIDpKol].Value.ToString());
+            }
+            if (what.Length != 0)
             {
                 var fromAdress = new MailAddress("aquadromautomat@gmail.com", "System automatycznej informacji");
                 var toAdress = new MailAddress("aquadromboss@gmail.com", "Administrator Firmy sMMonpar ");
                 const string fromPassword = "aquadrom123";
-                string subject = "Ostrzeżenie " + dataGridView1.Rows[iterator].Cells[Constants.PracownikImie].Value.ToString() + " " + dataGridView1.Rows[iterator].Cells[Constants.PracownikNazwisko].Value.ToString();
-
-                if (changeKPP == true) what = " Data ważności KPP: " + KPPdate.ToString("dd-MM-yyyy");
-                else what = " Data ważności badań lekarskich " + medicaldate.ToString("dd-MM-yyyy");
-
-                string body = "Pracownik: " + dataGridView1.Rows[iterator].Cells[Constants.PracownikImie].Value.ToString() + " " + dataGridView1.Rows[iterator].Cells[Constants.PracownikNazwisko].Value.ToString() + "\n" + what;
+                string subject = "Ostrzeżenie " + dataGridView1.Rows[iterator].Cells[Constants.PracownikImieKol].Value.ToString() + " " + dataGridView1.Rows[iterator].Cells[Constants.PracownikNazwiskoKol].Value.ToString();
+                string body = "Pracownik: " + dataGridView1.Rows[iterator].Cells[Constants.PracownikImieKol].Value.ToString() + " " + dataGridView1.Rows[iterator].Cells[Constants.PracownikNazwiskoKol].Value.ToString() + "\n" + what;
                 var smtp = new SmtpClient
                 {
                     Host = "smtp.gmail.com",
@@ -139,8 +169,12 @@ namespace aquadrom
                 {
                     smtp.Send(message);
                 }
-                adapter.Update(Constants.TabPracownik + " set " + Constants.PracownikOstrzezenie + "='t' where " + Constants.PracownikID + "=" + dataGridView1.Rows[iterator].Cells[Constants.PracownikID].Value.ToString());
             }
+        }
+
+        private void dataGridView1_Sorted(object sender, EventArgs e)
+        {
+            ColorCheckUser();
         }
     }
 }
